chore: setup demo
