+++ apps/benchmark/src/duck/switch/duck.tsx
@@ -1,20 +1,11 @@
-import { RadioGroup, RadioGroupItem } from '@gentleduck/registry-ui-duckui/radio-group'
+import { Switch } from '@gentleduck/registry-ui-duckui/switch'
 
-export default function RadioGroupDemo() {
+export default function SwitchDemo() {
   return (
-    <RadioGroup defaultValue="comfortable">
+    <div>
       <div className="flex items-center space-x-2">
-        <RadioGroupItem value="default" id="r1" />
-        <label htmlFor="r1">Default</label>
+        <Switch id="terms" />
       </div>
-      <div className="flex items-center space-x-2">
-        <RadioGroupItem value="comfortable" id="r2" />
-        <label htmlFor="r2">Comfortable</label>
-      </div>
-      <div className="flex items-center space-x-2">
-        <RadioGroupItem value="compact" id="r3" />
-        <label htmlFor="r3">Compact</label>
-      </div>
-    </RadioGroup>
+    </div>
   )
 }
