+++ packages/registry-ui-duckui/src/checkbox/index.ts
@@ -1,1 +1,1 @@
-export * from './checkbox'
+export * from './_checkbox'
