+++ packages/registry-ui-duckui/src/radio-group/index.ts
@@ -1,3 +1,3 @@
-export * from './radio-group'
+export * from './_radio-group'
 export * from './radio-group.hooks'
 export * from './radio-group.types'
